saving a workout works nowgit add . Im sure all my tests are failing

+++ back-end/src/main/java/com/allyoucanexercise/back_end/exerciseSet/ExerciseSetService.java
@@ -27,7 +27,7 @@ public class ExerciseSetService {
     }
 
     public ExerciseSet saveExerciseSet(WorkoutExercise workoutExercise, Integer setOrder, Integer reps, Float weight,
-            Integer durationSeconds, Integer distanceMeters) {
+            Integer durationSeconds, Float distanceMeters) {
         ExerciseSet exerciseSet = new ExerciseSet();
         exerciseSet.setWorkoutExercise(workoutExercise);
         exerciseSet.setSetOrder(setOrder);
